requirement.txt
pytest
selenium
jsonpickle
pony
pymysql
pytest-allure-adaptor
